--- charge/src/main/java/de/elvah/charge/platform/network/retrofit/ApiKeyInterceptor.kt
@@ -1,5 +1,0 @@
-package de.elvah.charge.platform.network.retrofit
-
-import okhttp3.Interceptor
-
-internal interface ApiKeyInterceptor : Interceptor
